requirement.txt
asgiref==3.2.3
Django==3.0
djangorestframework==3.11.0
numpy==1.17.4
pandas==0.25.3
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
